Disable Expect Header prior to POST
The issue is that the .NET client set the expect header and only send the request headers before a POST of data. This allows the server to respond with errors/redirects/security violations prior to the client sending the request body. The client does not wait until it gets a response and just pushes out the body of the request, which results in a 417 expectation error on my apache 2.2 and requesting auth token failed.